requirements.txt
torch>=1.9.0
numpy>=1.19.0
matplotlib>=3.3.0
plotly>=5.0.0      
pyyaml>=5.4.0      
tensorboard>=2.7.0  
scikit-learn>=0.24.0   
scipy>=1.7.0        
torch-geometric>=2.0.0
gudhi>=3.5.0          
tqdm>=4.62.0       